add send/recv helpers; rework response headers
The 'send' and (the less important) 'recv' functions responsible for
ensuring proper line termination when sending data to the client.  The
response headers are now stored in an array.  Remove 'set -x', as it is
really noisy, and only useful for debugging purposes. 'send' and 'recv'
will trace what is sent to stdout/read from stdin to accomplish similar
debugging purposes.